+++ flutter_inappwebview/lib/src/in_app_browser/in_app_browser.dart
@@ -412,9 +412,6 @@ class InAppBrowser implements PlatformInAppBrowserEvents {
   ) {}
 
   @override
-  void onFaviconChanged(FaviconChangedRequest faviconChangedRequest) {}
-
-  @override
   void onReceivedLoginRequest(LoginRequest loginRequest) {}
 
   @override