+++ flutter_inappwebview/lib/src/in_app_webview/headless_in_app_webview.dart
@@ -225,11 +225,6 @@ class HeadlessInAppWebView {
       PermissionRequest permissionRequest,
     )?
     onPermissionRequest,
-    void Function(
-      InAppWebViewController controller,
-      FaviconChangedRequest faviconChangedRequest,
-    )?
-    onFaviconChanged,
     void Function(InAppWebViewController controller, LoginRequest loginRequest)?
     onReceivedLoginRequest,
     void Function(
@@ -482,10 +477,6 @@ class HeadlessInAppWebView {
                : null,
            onFormResubmission: onFormResubmission != null
                ? (controller, url) => onFormResubmission.call(controller, url)
-               : null,
-           onFaviconChanged: onFaviconChanged != null
-               ? (controller, request) =>
-                     onFaviconChanged.call(controller, request)
                : null,
            onReceivedTouchIconUrl: onReceivedTouchIconUrl != null
                ? (controller, url, precomposed) =>