+++ flutter_inappwebview_platform_interface/lib/src/in_app_webview/platform_headless_in_app_webview.dart
@@ -72,7 +72,6 @@ class PlatformHeadlessInAppWebViewCreationParams
     super.onRenderProcessResponsive,
     super.onRenderProcessUnresponsive,
     super.onFormResubmission,
-    super.onFaviconChanged,
     super.onReceivedTouchIconUrl,
     super.onJsBeforeUnload,
     super.onReceivedLoginRequest,