+++ flutter_inappwebview_platform_interface/lib/src/in_app_webview/platform_inappwebview_widget.dart
@@ -45,7 +45,6 @@ class PlatformInAppWebViewWidgetCreationParams
     super.onConsoleMessage,
     super.shouldOverrideUrlLoading,
     super.onLoadResource,
-    super.onFaviconChanged,
     super.onScrollChanged,
     super.onDownloadStarting,
     super.onLoadResourceWithCustomScheme,