+++ flutter_inappwebview_platform_interface/lib/src/types/main.dart
@@ -37,7 +37,6 @@ export 'security_restriction_mode.dart' show SecurityRestrictionMode;
 export 'dismiss_button_style.dart' show DismissButtonStyle;
 export 'download_start_request.dart' show DownloadStartRequest;
 export 'favicon.dart' show Favicon;
-export 'favicon_changed_request.dart' show FaviconChangedRequest;
 export 'fetch_request.dart' show FetchRequest;
 export 'fetch_request_action.dart' show FetchRequestAction;
 export 'fetch_request_credential.dart' show FetchRequestCredential;